requirements.txt
streamlit>=1.32.0
groq>=0.4.2
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-core>=0.1.0
python-dotenv>=1.0.0
sounddevice>=0.4.6
numpy==1.26.4
scipy>=1.11.3
SpeechRecognition>=3.10.0
wave>=0.0.2
pyaudio>=0.2.13
transformers>=4.38.2
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
pandas>=2.0.0
openai-whisper>=20231117
validators>=0.22.0
pytubefix>=1.11.2
gspread>=5.12.0
google-auth>=2.23.0
ffmpeg-python>=0.2.0
sympy==1.13.1